:construction: Turning cache into module
