+++ src/components/Patterns.vue
@@ -42,6 +42,7 @@ export default class Patterns extends Vue {
 
   &:hover
     box-shadow: inset 0 0 100px 100px rgba(255, 255, 255, 0.1)
+    cursor: pointer
 
 .selected
   border: 1px solid rgba(255, 255, 255, 0.36)
