+++ src/schemas.ts
@@ -1,8 +1,7 @@
 import { autoserialize, autoserializeAs } from 'cerialize';
 import Tone from 'tone';
 import Part from './modules/audio/part';
-import { VuexModule, Mutation } from 'vuex-module-decorators';
-import { Module } from 'vuex';
+
 
 export interface INote {
   id: number;
@@ -122,50 +121,5 @@ export class Instrument implements IInstrument {
   }
   public triggerAttackRelease(note: string, duration: string, time: number) {
     this.synth.triggerAttackRelease(note, duration, time);
-  }
-}
-
-export class Project extends VuexModule {
-  @autoserialize public bpm = 128;
-  @autoserializeAs(Pattern) public patterns: Pattern[] = [];
-  @autoserializeAs(Instrument) public instruments: Instrument[] = [];
-  constructor(module?: Module<any, any>) {
-    super(module || {});
-  }
-
-  @Mutation
-  public reset(payload: Project) {
-    Object.assign(this, payload);
-  }
-
-  @Mutation
-  public addInstrument() {
-    let name: string;
-    let count = 0;
-    while (true) {
-      name = `Instrument ${count}`;
-      let found = false;
-      for (const instrument of this.instruments) {
-        if (instrument.name === name) {
-          found = true;
-          break;
-        }
-      }
-
-      if (!found) {
-        break;
-      }
-
-      count++;
-    }
-    this.instruments.push(Instrument.default(name));
-  }
-
-  get patternLookup() {
-    const patterns: {[k: string]: Pattern} = {};
-    this.patterns.forEach((pattern) => {
-      patterns[pattern.name] = pattern;
-    });
-    return patterns;
   }
 }
