+++ src/store/general.ts
@@ -1,0 +1,10 @@
+import { VuexModule } from 'vuex-module-decorators';
+
+import SideBar from '@/components/SideBar.vue';
+import BaseTabs from '@/components/BaseTabs.vue';
+
+export default class General extends VuexModule {
+  public sideBarTabs: SideBar[] = [];
+  public panels: BaseTabs | null = null;
+  public toolbarHeight = 64;
+}
