--- src/views/Main.vue
@@ -1,15 +1,0 @@
-<template>
-  <div></div>
-</template>
-
-<script lang="ts">
-import Vue from 'vue';
-import { Component } from 'vue-property-decorator';
-
-@Component
-export default class Main extends Vue {}
-</script>
-
-<style scoped>
-
-</style>
